% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{addMissing}
\alias{addMissing}
\title{Simulate missing values with mimimal assumptions}
\usage{
addMissing(obj, n = Inf, id_column, verbose = TRUE)
}
\arguments{
\item{obj}{Input MSnSet}

\item{n}{Number of PSMs to impute missing values in}

\item{id_column}{Column to group PSM by}

\item{verbose}{Output description of missing values to console}
}
\value{
list("truth":Input data set with ground truths for missing values,
             "missing":Input data set with NA for missing values)
}
\description{
Assumes that PSMs from the same peptide should have the same
relative abundances, and that the missing values are missing at random (MAR)
with peptide intensity the main cause of missingness.
}
\details{
Identifies groups of PSMs with at least one with missing values (PSMm) and
one without (PSMnm) from the same peptide. PSMnm are then randomly paired
with a PSMm, intensity values scaled to the PSMm and missing values
(\code{NA}) added at the same positions as missing values in PSMm.

Returns two MSnSets: "truth" and "missing". Both have the same dimensions as
the input MSnSet. Truth contains the scaled PSMs. Missing contains the scaled
PSMs with missing values added.
}
