% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{bn}
\alias{bn}
\title{Adjust quantification values using bridging channels}
\usage{
bn(values, tag1_ix, tag2_ix, bridge1_ix, bridge2_ix)
}
\arguments{
\item{values}{vector of quantification values}

\item{tag1_ix}{index for quantification values in multiplex 1}

\item{tag2_ix}{index for quantification values in multiplex 2}

\item{bridge1_ix}{index for bridge sample(s) in multiplex 1}

\item{bridge2_ix}{index for bridge sample(s) in multiplex 1}
}
\value{
bridge normalised values
}
\description{
Adjust quantification values using bridging channels to
normalise across isobaric tag multiplexes.
Note that the bridge channels are not normalised.
}
\examples{
bn(c(1,1,2,2,2,4), tag1_ix=1:2, tag2_ix=4:5, bridge1_ix=3, bridge2_ix=6)
}
