% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{bridgeNormalise}
\alias{bridgeNormalise}
\title{Adjust quantification values in MSnSet}
\usage{
bridgeNormalise(obj, tag1_ix = 1:10, tag2_ix = 12:21,
  bridge1_ix = 11, bridge2_ix = 22, keep_bridge = "one")
}
\arguments{
\item{tag1_ix}{index for quantification values in multiplex 1}

\item{tag2_ix}{index for quantification values in multiplex 2}

\item{bridge1_ix}{index for bridge sample(s) in multiplex 1}

\item{bridge2_ix}{index for bridge sample(s) in multiplex 1}

\item{keep_bridge}{'both': keep both bridge samples. 'one': keep just one.}

\item{values}{vector of quantification values}
}
\value{
bridge normalised values
}
\description{
Adjust quantification values in MSnSet using bridging channels to
normalise across isobaric tag multiplexes.
Note that the bridge channels are not normalised.
}
