% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combineTMTsets}
\alias{combineTMTsets}
\title{Combine TMT sets.}
\usage{
combineTMTsets(x, y)
}
\arguments{
\item{x}{MSnSet containing quanification values for TMT set 1}

\item{y}{MSnSet containing quanification values for TMT set 2}
}
\value{
Combined MsnSet
}
\description{
Note that only intersecting features are returned
}
