% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{getMAQE}
\alias{getMAQE}
\title{Obtain the Mean Absolute Quotient Error (MAQE)}
\usage{
getMAQE(truth, imputed)
}
\arguments{
\item{truth}{ground truths}

\item{imputed}{imputed values}
}
\value{
MAQE
}
\description{
Obtain the Mean Absolute Quotient Error (MAQE)
}
