% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{getMARD}
\alias{getMARD}
\title{Obtain the Mean Absolute Relative Difference (MARD)}
\usage{
getMARD(truth, imputed)
}
\arguments{
\item{truth}{ground truths}

\item{imputed}{imputed values}
}
\value{
MARD
}
\description{
Obtain the Mean Absolute Relative Difference (MARD)
}
