% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM.R
\name{getOrganelleF1}
\alias{getOrganelleF1}
\title{Plot the missing values.}
\usage{
getOrganelleF1(params)
}
\arguments{
\item{params}{Output from \code{svmOptimisation}}
}
\value{
\code{data.frame()} with per organelle F1
}
\description{
Plot the missing values.
}
