% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsep.R
\name{getQsepDistances}
\alias{getQsepDistances}
\title{Obtain the QSep distances for an MSnSet with marker proteins annotated}
\usage{
getQsepDistances(obj, ...)
}
\arguments{
\item{obj}{MSnSet input}

\item{...}{further arguments passed to QSep}
}
\value{
\code{data.frame()} with qsep distances
}
\description{
Obtain the QSep distances for an MSnSet with marker proteins annotated
}
