% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{getRMedSE}
\alias{getRMedSE}
\title{Obtain the Root median squared error (RMedSE)}
\usage{
getRMedSE(truth, imputed)
}
\arguments{
\item{truth}{ground truths}

\item{imputed}{imputed values}
}
\value{
RMedSE
}
\description{
Obtain the Root median squared error (RMedSE)
}
