% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{getRMSE}
\alias{getRMSE}
\title{Obtain the Root mean squared error (RMSE)}
\usage{
getRMSE(truth, imputed)
}
\arguments{
\item{truth}{ground truths}

\item{imputed}{imputed values}
}
\value{
RMSE
}
\description{
Obtain the Root mean squared error (RMSE)
}
