% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{getTruthImputed}
\alias{getTruthImputed}
\title{Extract the truth and imputed values}
\usage{
getTruthImputed(obj, missing_obj, truth_obj)
}
\arguments{
\item{obj}{MSnSet with imputed values}

\item{missing_obj}{MSnSet with missing values}

\item{truth_obj}{MSnSet with ground truths}
}
\value{
data.frame() with truth and imputed columns
}
\description{
Extract the truth and imputed values
}
