% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{imputeOptProc}
\alias{imputeOptProc}
\title{Impute missing values using either \code{\link{MSnbase::impute}} or
\code{\link{RobCompositions}}}
\usage{
imputeOptProc(obj, method, k, verbose = FALSE)
}
\arguments{
\item{obj}{MSnSet with missing values}

\item{method}{imputation method, Choose from knn, MinDet, MinProb,
sn-knn, max-knn, it-comp-knn or comp-knn}
}
\value{
input MSnSet with missing values imputed
}
\description{
Impute missing values using either \code{\link{MSnbase::impute}} or
\code{\link{RobCompositions}}
}
\seealso{
\code{\link{robCompositions::impKNNa}} for comp-knn and
\code{\link{robCompositions::impCoda}} for it-comp-knn methods
}
