% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{mb}
\alias{mb}
\title{Average the bridge sample(s) in a vector of quantification values}
\usage{
mb(values, plex1_ix, plex2_ix, bridge1_ix, bridge2_ix)
}
\arguments{
\item{values}{vector of quantification values}

\item{plex1_ix}{index for quantification values in multiplex 1}

\item{plex2_ix}{index for quantification values in multiplex 2}

\item{bridge1_ix}{index for bridge sample(s) in multiplex 1}

\item{bridge2_ix}{index for bridge sample(s) in multiplex 1}
}
\value{
bridge normalised values
}
\description{
Note, redundant bridge sample(s) are removed
}
\examples{
q <- c(1,1,NA,3,2,2,4,6)
MeanBridges(q, plex1_ix=1:2, plex2_ix=5:6, bridge1_ix=3:4, bridge2_ix=7:8)
}
