% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{MeanBridges}
\alias{MeanBridges}
\title{Average the bridge sample(s) for each feature in an MSnSet}
\usage{
MeanBridges(obj, plex1_ix, plex2_ix, bridge1_ix, bridge2_ix)
}
\arguments{
\item{plex1_ix}{index for quantification values in multiplex 1}

\item{plex2_ix}{index for quantification values in multiplex 2}

\item{bridge1_ix}{index for bridge sample(s) in multiplex 1}

\item{bridge2_ix}{index for bridge sample(s) in multiplex 1}

\item{values}{vector of quantification values}
}
\value{
bridge normalised values
}
\description{
Note, redundant bridge sample(s) are removed
}
