% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{MeanBridgesMsnSet}
\alias{MeanBridgesMsnSet}
\title{Average the bridge sample(s) for each feature in an MSnSet}
\usage{
MeanBridgesMsnSet(obj, tag1_ix, tag2_ix, bridge1_ix, bridge2_ix)
}
\arguments{
\item{tag1_ix}{index for quantification values in multiplex 1}

\item{tag2_ix}{index for quantification values in multiplex 2}

\item{bridge1_ix}{index for bridge sample(s) in multiplex 1}

\item{bridge2_ix}{index for bridge sample(s) in multiplex 1}

\item{values}{vector of quantification values}
}
\value{
bridge normalised values
}
\description{
Note, redundant bridge sample(s) are removed
}
