% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{plotMissing}
\alias{plotMissing}
\title{Plot the missing values.}
\usage{
plotMissing(obj, verbose = TRUE, ...)
}
\arguments{
\item{obj}{Input}

\item{verbose}{Tally the missing values}

\item{...}{further arguments passed to gplots::heatmap.2}
}
\value{
NA
}
\description{
Plot the missing values.
}
\seealso{
\code{\link{heatmap.2}} which this function wraps
}
