% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsep.R
\name{plotQsepDistances}
\alias{plotQsepDistances}
\title{Plot the QSep distances obtained from \code{getQsepDistances}}
\usage{
plotQsepDistances(qsep_df)
}
\arguments{
\item{qsep_df}{Output from \code{getQsepDistances}}
}
\value{
\code{ggplot2} plot
}
\description{
Plot the QSep distances obtained from \code{getQsepDistances}
}
