% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{plotTruthImputed}
\alias{plotTruthImputed}
\title{Plot the truth vs imputed}
\usage{
plotTruthImputed(obj)
}
\arguments{
\item{obj}{data.frame() with truth and imputed columns}
}
\value{
scatter plot for truth vs imputed
}
\description{
Plot the truth vs imputed
}
