% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{robComp}
\alias{robComp}
\title{Impute missing values using compositional KNN imputation}
\usage{
robComp(obj, method = "comp-knn", k, sum_first = FALSE, verbose = FALSE)
}
\arguments{
\item{obj}{Input}

\item{method}{impKNNa=compositional KNN or
impCoda=compositional KNN with iterative improvement in estimates}

\item{k}{nearest neighbours}

\item{sum_first}{Sum normalise the array data before imputation, and convert
back to non-sum normalised afterwards}

\item{verbose}{Log the start and end time for imputation}
}
\value{
Input data with missing values imputed
}
\description{
Impute missing values using compositional KNN imputation
}
\seealso{
\code{\link{robCompositions::impKNNa}} and
\code{\link{robCompositions::impCoda}} which this function wraps
}
