% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{spatial_proteomics_normalise}
\alias{spatial_proteomics_normalise}
\title{Normalise data using sum +/- VSN}
\usage{
spatial_proteomics_normalise(obj, method = "sum")
}
\arguments{
\item{obj}{Input}

\item{method}{sum = row-wise sum normalisation, or
vsn-sum = VSN then row-wise sum normalisation, or
sum-vsn = row-wise sum normalisation then VSN}
}
\value{
Normalised Input data
}
\description{
Normalise data using sum +/- VSN
}
