% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{getRMSE}
\alias{getRMSE}
\title{Obtain the Root mean squared error (RMSE)}
\usage{
getRMSE(obj)
}
\arguments{
\item{obj}{data.frame() with truth and imputed columns}
}
\value{
RMSE
}
\description{
Obtain the Root mean squared error (RMSE)
}
